Refactor: Extract sub-ViewModels from MainViewModel.cs
Separated BaseElementViewModel and ObservableObject into ViewModels/Base.
Extracted ViewModels for Track, Signal, Switch, Route, Infrastructure, etc. into ViewModels/Elements.
Updated MainWindow and XAML namespaces.
