+++ RailmlEditor/Logic/TopologyManager.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using RailmlEditor.ViewModels;
+using RailmlEditor.ViewModels.Elements;
 
 namespace RailmlEditor.Logic
 {
@@ -301,3 +302,5 @@ namespace RailmlEditor.Logic
         }
     }
 }
+
+
