+++ RailmlEditor/MainWindow.xaml.cs
@@ -4,6 +4,7 @@ using System.Windows;
 using System.Windows.Controls;
 using System.Windows.Input;
 using RailmlEditor.ViewModels;
+using RailmlEditor.ViewModels.Elements;
 using System.Windows.Media;
 
 namespace RailmlEditor
@@ -1282,3 +1283,5 @@ namespace RailmlEditor
         }
     }
 }
+
+
