+++ RailmlEditor/Utils/UndoRedoManager.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using RailmlEditor.ViewModels;
+using RailmlEditor.ViewModels.Elements;
 
 namespace RailmlEditor.Utils
 {
@@ -70,4 +71,6 @@ namespace RailmlEditor.Utils
         }
     }
 }
+
+
