+++ RailmlEditor/ViewModels/Base/BaseElementViewModel.cs
@@ -1,0 +1,55 @@
+using System;
+using System.ComponentModel;
+using RailmlEditor.Utils;
+
+namespace RailmlEditor.ViewModels
+{
+    public abstract class BaseElementViewModel : ObservableObject
+    {
+        private double _x;
+        private double _y;
+        private string? _id;
+        private bool _isSelected;
+
+        public virtual double X
+        {
+            get => _x;
+            set => SetProperty(ref _x, value);
+        }
+
+        public virtual double Y
+        {
+            get => _y;
+            set => SetProperty(ref _y, value);
+        }
+
+        public string? Id
+        {
+            get => _id;
+            set => SetProperty(ref _id, value);
+        }
+
+        private string? _name;
+        public string? Name
+        {
+            get => _name;
+            set => SetProperty(ref _name, value);
+        }
+
+        public bool IsSelected
+        {
+            get => _isSelected;
+            set => SetProperty(ref _isSelected, value);
+        }
+
+        private bool _showCoordinates;
+        public bool ShowCoordinates
+        {
+            get => _showCoordinates;
+            set => SetProperty(ref _showCoordinates, value);
+        }
+
+        public abstract string TypeName { get; }
+    }
+}
+
