+++ RailmlEditor/ViewModels/Base/ObservableObject.cs
