+++ RailmlEditor/ViewModels/Elements/SignalViewModel.cs
@@ -1,0 +1,71 @@
+using System.Collections.Generic;
+
+namespace RailmlEditor.ViewModels.Elements
+{
+    public class SignalViewModel : BaseElementViewModel
+    {
+        public override string TypeName => "Signal";
+        
+        // Collections for ComboBox
+        public System.Collections.Generic.List<string> AvailableTypes { get; } = new System.Collections.Generic.List<string> 
+        { 
+            "main", "distant", "repeater", "shunting" 
+        };
+
+        public System.Collections.Generic.List<string> AvailableFunctions { get; } = new System.Collections.Generic.List<string> 
+        { 
+            "exit", "home", "blocking" 
+        };
+
+        private string? _type;
+        public string? Type
+        {
+            get => _type;
+            set => SetProperty(ref _type, value);
+        }
+
+        private string? _function;
+        public string? Function
+        {
+            get => _function;
+            set => SetProperty(ref _function, value);
+        }
+        
+        private string? _relatedTrackId;
+        public string? RelatedTrackId
+        {
+            get => _relatedTrackId;
+            set => SetProperty(ref _relatedTrackId, value);
+        }
+
+        private double _pos;
+        public double Pos
+        {
+            get => _pos;
+            set => SetProperty(ref _pos, value);
+        }
+
+        public System.Collections.Generic.List<string> AvailableDirections { get; } = new System.Collections.Generic.List<string> { "up", "down" };
+        
+        private string _direction = "up";
+        public string Direction
+        {
+            get => _direction;
+            set 
+            {
+                if (SetProperty(ref _direction, value))
+                {
+                    OnPropertyChanged(nameof(IsFlipped));
+                }
+            }
+        }
+
+        public bool IsFlipped
+        {
+            get => _direction == "down";
+            set => Direction = value ? "down" : "up";
+        }
+    }
+}
+
+
