+++ RailmlEditor/ViewModels/MainViewModel.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.Collections.Specialized;
@@ -7,675 +8,12 @@ using System.Windows;
 using System.Windows.Input;
 using RailmlEditor.Models;
 using RailmlEditor.Utils;
+using RailmlEditor.ViewModels.Elements;
 using System.IO;
 using System.Text.Json;
 
 namespace RailmlEditor.ViewModels
 {
-    public abstract class BaseElementViewModel : ObservableObject
-    {
-        private double _x;
-        private double _y;
-        private string? _id;
-        private bool _isSelected;
-
-        public virtual double X
-        {
-            get => _x;
-            set => SetProperty(ref _x, value);
-        }
-
-        public virtual double Y
-        {
-            get => _y;
-            set => SetProperty(ref _y, value);
-        }
-
-        public string? Id
-        {
-            get => _id;
-            set => SetProperty(ref _id, value);
-        }
-
-        private string? _name;
-        public string? Name
-        {
-            get => _name;
-            set => SetProperty(ref _name, value);
-        }
-
-        public bool IsSelected
-        {
-            get => _isSelected;
-            set => SetProperty(ref _isSelected, value);
-        }
-
-        private bool _showCoordinates;
-        public bool ShowCoordinates
-        {
-            get => _showCoordinates;
-            set => SetProperty(ref _showCoordinates, value);
-        }
-
-        public abstract string TypeName { get; }
-    }
-
-    public enum TrackNodeType
-    {
-        None,
-        BufferStop,
-        OpenEnd,
-        Connection
-    }
-
-    public class TrackViewModel : BaseElementViewModel
-    {
-        public override string TypeName => "Track";
-        
-        // Collections for ComboBox
-        public System.Collections.Generic.List<string> AvailableTypes { get; } = new System.Collections.Generic.List<string> 
-        { 
-            "mainTrack", "secondaryTrack", "connectingTrack", "sidingTrack", "stationTrack" 
-        };
-
-        public bool IsCurved => this is CurvedTrackViewModel;
-        
-
-        public System.Collections.Generic.List<string> AvailableMainDirs { get; } = new System.Collections.Generic.List<string> 
-        { 
-            "up", "down", "none" 
-        };
-
-        public ICommand FlipHorizontallyCommand { get; }
-        public ICommand FlipVerticallyCommand { get; }
-
-        public TrackViewModel()
-        {
-            FlipHorizontallyCommand = new RelayCommand(_ => FlipHorizontally());
-            FlipVerticallyCommand = new RelayCommand(_ => FlipVertically());
-            // This is TrackViewModel... oops. I need to edit MainViewModel!
-            // Wait, I am editing MainViewModel.cs, but scrolling shows TrackViewModel constructor.
-            // I need to find MainViewModel constructor.
-            // Initialize Nodes
-            BeginNode.Role = "Begin";
-            BeginNode.Pos = 0;
-            EndNode.Role = "End";
-            EndNode.Pos = Length;
-            Children.Add(BeginNode);
-            Children.Add(EndNode);
-
-            BeginNode.PropertyChanged += OnNodePropertyChanged;
-            EndNode.PropertyChanged += OnNodePropertyChanged;
-            this.PropertyChanged += OnTrackPropertyChanged;
-        }
-
-        private void OnTrackPropertyChanged(object? sender, PropertyChangedEventArgs e)
-        {
-            if (e.PropertyName == nameof(Id))
-            {
-                if (!string.IsNullOrEmpty(Id))
-                {
-                    string num = System.Text.RegularExpressions.Regex.Match(Id, @"\d+").Value;
-                    BeginNode.Id = "tb" + num;
-                    EndNode.Id = "te" + num;
-                    BeginNode.ConnectionId = "cb" + num;
-                    EndNode.ConnectionId = "ce" + num;
-                }
-            }
-            // Trigger connection check on geometry change
-            if (e.PropertyName == nameof(X) || e.PropertyName == nameof(Y) || 
-                e.PropertyName == nameof(X2) || e.PropertyName == nameof(Y2) ||
-                e.PropertyName == nameof(CurvedTrackViewModel.MX) || e.PropertyName == nameof(CurvedTrackViewModel.MY))
-            {
-                // We need access to the MainViewModel instance or Elements collection to check others.
-                // TrackViewModel doesn't know about MainViewModel.
-                // We must handle this in MainViewModel.Elements_CollectionChanged -> Item.PropertyChanged (already there!)
-            }
-            if (e.PropertyName == nameof(Length))
-            {
-                EndNode.Pos = Length;
-            }
-        }
-
-        private void OnNodePropertyChanged(object? sender, System.ComponentModel.PropertyChangedEventArgs e)
-        {
-            if (e.PropertyName == nameof(TrackNodeViewModel.NodeType))
-            {
-                if (sender == BeginNode && BeginType != BeginNode.NodeType)
-                {
-                    BeginType = BeginNode.NodeType;
-                }
-                else if (sender == EndNode && EndType != EndNode.NodeType)
-                {
-                    EndType = EndNode.NodeType;
-                }
-                
-                // Update ConnectionId when NodeType changes
-                if (!string.IsNullOrEmpty(Id))
-                {
-                    string num = System.Text.RegularExpressions.Regex.Match(Id, @"\d+").Value;
-                    if (sender == BeginNode) BeginNode.ConnectionId = "cb" + num;
-                    else if (sender == EndNode) EndNode.ConnectionId = "ce" + num;
-                }
-            }
-            if (e.PropertyName == nameof(TrackNodeViewModel.ConnectedNodeId) || e.PropertyName == nameof(TrackNodeViewModel.ConnectedTrackId))
-            {
-                var node = sender as TrackNodeViewModel;
-                if (node != null)
-                {
-                    node.ConnectionRef = node.ConnectedNodeId;
-                }
-            }
-        }
-
-        protected virtual void FlipHorizontally()
-        {
-            double tempX = X;
-            X = X2;
-            X2 = tempX;
-            OnPropertyChanged(nameof(X));
-            OnPropertyChanged(nameof(X2));
-            OnPropertyChanged(nameof(Length));
-        }
-
-        protected virtual void FlipVertically()
-        {
-            // By default, do nothing or only for curved tracks if required.
-            // But user said: "Flip Vertically only for code=corner"
-        }
-
-        private string? _description;
-        public string? Description
-        {
-            get => _description;
-            set => SetProperty(ref _description, value);
-        }
-
-        private string? _type;
-        public string? Type
-        {
-            get => _type;
-            set => SetProperty(ref _type, value);
-        }
-
-        private string? _mainDir = "up";
-        public string? MainDir
-        {
-            get => _mainDir;
-            set => SetProperty(ref _mainDir, value);
-        }
-
-        private string? _code;
-        public string? Code
-        {
-            get => _code;
-            set => SetProperty(ref _code, value);
-        }
-        
-        public override double X
-        {
-            get => base.X;
-            set
-            {
-                if (base.X != value)
-                {
-                    double delta = value - base.X;
-                    base.X = value;
-                    OnPropertyChanged(nameof(X));
-                    OnPropertyChanged(nameof(Length));
-
-                    // Move Child Signals and Borders
-                    if (Children != null)
-                    {
-                        foreach (var child in Children)
-                        {
-                            if (child is SignalViewModel signal)
-                            {
-                                signal.X += delta;
-                            }
-                            else if (child is TrackCircuitBorderViewModel border)
-                            {
-                                border.X += delta;
-                            }
-                        }
-                        UpdateBorderAngles();
-                    }
-                }
-            }
-        }
-
-        public override double Y
-        {
-            get => base.Y;
-            set
-            {
-                if (base.Y != value)
-                {
-                    double delta = value - base.Y;
-                    base.Y = value;
-                    OnPropertyChanged(nameof(Y));
-                    OnPropertyChanged(nameof(Length));
-
-                    // Move Child Signals and Borders
-                    if (Children != null)
-                    {
-                        foreach (var child in Children)
-                        {
-                            if (child is SignalViewModel signal)
-                            {
-                                signal.Y += delta;
-                            }
-                            else if (child is TrackCircuitBorderViewModel border)
-                            {
-                                border.Y += delta;
-                            }
-                        }
-                        UpdateBorderAngles();
-                    }
-                }
-            }
-        }
-
-        private double _x2;
-        public virtual double X2
-        {
-            get => _x2;
-            set 
-            {
-                if (SetProperty(ref _x2, value))
-                {
-                    OnPropertyChanged(nameof(Length));
-                    UpdateBorderAngles();
-                }
-            }
-        }
-
-        private double _y2;
-        public virtual double Y2
-        {
-            get => _y2;
-            set 
-            {
-                if (SetProperty(ref _y2, value))
-                {
-                    OnPropertyChanged(nameof(Length));
-                    UpdateBorderAngles();
-                }
-            }
-        }
-
-        private void UpdateBorderAngles()
-        {
-            if (Children == null) return;
-            double angle = Math.Atan2(Y2 - Y, X2 - X) * 180 / Math.PI;
-            foreach (var border in Children.OfType<TrackCircuitBorderViewModel>())
-            {
-                border.Angle = angle;
-            }
-        }
-
-        // Length is now derived or updates X2?
-        // Let's make Length read-only derived, or if set, it updates X2 (assuming horizontal extension).
-        public virtual double Length
-        {
-            get => System.Math.Sqrt(System.Math.Pow(X2 - X, 2) + System.Math.Pow(Y2 - Y, 2));
-            set
-            {
-                // If setting length, assume extending horizontally from X
-                X2 = X + value;
-                Y2 = Y;
-                OnPropertyChanged(nameof(Length));
-            }
-        }
-
-        public ObservableCollection<BaseElementViewModel> Children { get; } = new();
-
-        public TrackNodeViewModel BeginNode { get; } = new TrackNodeViewModel();
-        public TrackNodeViewModel EndNode { get; } = new TrackNodeViewModel();
-
-        // Begin/End Configuration
-        private TrackNodeType _beginType;
-        public TrackNodeType BeginType
-        {
-            get => _beginType;
-            set
-            {
-                if (SetProperty(ref _beginType, value))
-                {
-                    OnBeginTypeChanged();
-                }
-            }
-        }
-
-        private TrackNodeType _endType;
-        public TrackNodeType EndType
-        {
-            get => _endType;
-            set
-            {
-                if (SetProperty(ref _endType, value))
-                {
-                    OnEndTypeChanged();
-                }
-            }
-        }
-
-        // Connection State (to disable fields)
-        private bool _hasBeginConnection;
-        public bool HasBeginConnection
-        {
-            get => _hasBeginConnection;
-            set
-            {
-                if (SetProperty(ref _hasBeginConnection, value))
-                {
-                    OnPropertyChanged(nameof(IsBeginEnabled));
-                }
-            }
-        }
-
-        private bool _hasEndConnection;
-        public bool HasEndConnection
-        {
-            get => _hasEndConnection;
-            set
-            {
-                if (SetProperty(ref _hasEndConnection, value))
-                {
-                    OnPropertyChanged(nameof(IsEndEnabled));
-                }
-            }
-        }
-
-        public bool IsBeginEnabled => !HasBeginConnection;
-        public bool IsEndEnabled => !HasEndConnection;
-
-        // Auto-ID Logic & Children Management
-        private void OnBeginTypeChanged()
-        {
-            BeginNode.NodeType = BeginType;
-            BeginNode.Role = "Begin";
-            
-            if (!Children.Contains(BeginNode)) Children.Insert(0, BeginNode); // Insert at top
-        }
-
-        private void OnEndTypeChanged()
-        {
-            EndNode.NodeType = EndType;
-            EndNode.Role = "End";
-
-            if (!Children.Contains(EndNode)) Children.Add(EndNode); // Add at bottom (or after Begin)
-        }
-
-        private string GenerateId(string prefix)
-        {
-            // Extract number from Track ID
-            var match = System.Text.RegularExpressions.Regex.Match(Id ?? "", @"\d+$");
-            string number = match.Success ? match.Value : "1";
-            return $"{prefix}{number}";
-        }
-
-        public IEnumerable<TrackNodeType> AvailableTrackNodeTypes => System.Enum.GetValues(typeof(TrackNodeType)).Cast<TrackNodeType>();
-    }
-
-    public class SwitchViewModel : BaseElementViewModel
-    {
-        public override string TypeName => "Switch";
-        public double Pos { get; set; }
-
-        private string _trackContinueCourse = "straight";
-        public string TrackContinueCourse
-        {
-            get => _trackContinueCourse;
-            set => SetProperty(ref _trackContinueCourse, value);
-        }
-
-        private string _normalPosition = "straight";
-        public string NormalPosition
-        {
-            get => _normalPosition;
-            set => SetProperty(ref _normalPosition, value);
-        }
-
-        public ObservableCollection<DivergingConnectionViewModel> DivergingConnections { get; } = new();
-
-        public List<string> AvailableCourses { get; } = new() { "straight", "left", "right", "other" };
-
-        // Logic data
-        public string? PrincipleTrackId { get; set; }
-        public string? EnteringTrackId { get; set; }
-        public List<string> DivergingTrackIds { get; } = new();
-        public bool IsScenario1 { get; set; } // true: End -> multiple Begins, false: Begin -> multiple Ends
-    }
-
-    public class DivergingConnectionViewModel : ObservableObject
-    {
-        public string? TrackId { get; set; }
-
-        private string? _displayName;
-        public string? DisplayName
-        {
-            get => _displayName;
-            set => SetProperty(ref _displayName, value);
-        }
-
-        private string? _id;
-        public string? Id { get => _id; set => SetProperty(ref _id, value); }
-
-        private string? _ref;
-        public string? Ref { get => _ref; set => SetProperty(ref _ref, value); }
-
-        private string _orientation = "outgoing";
-        public string Orientation { get => _orientation; set => SetProperty(ref _orientation, value); }
-
-        private TrackViewModel? _targetTrack;
-        public TrackViewModel? TargetTrack
-        {
-            get => _targetTrack;
-            set
-            {
-                if (_targetTrack != null) _targetTrack.PropertyChanged -= Track_PropertyChanged;
-                _targetTrack = value;
-                if (_targetTrack != null)
-                {
-                    _targetTrack.PropertyChanged += Track_PropertyChanged;
-                    UpdateDisplayName();
-                }
-            }
-        }
-
-        private void Track_PropertyChanged(object? sender, PropertyChangedEventArgs e)
-        {
-            if (e.PropertyName == nameof(BaseElementViewModel.Name) || e.PropertyName == nameof(BaseElementViewModel.Id))
-            {
-                UpdateDisplayName();
-            }
-        }
-
-        private void UpdateDisplayName()
-        {
-            if (TargetTrack != null)
-            {
-                DisplayName = $"{TargetTrack.Id}({TargetTrack.Name ?? "unnamed"})";
-            }
-        }
-
-        private string _course = "straight";
-        public string Course
-        {
-            get => _course;
-            set => SetProperty(ref _course, value);
-        }
-
-        public List<string> AvailableCourses { get; } = new() { "straight", "left", "right" };
-    }
-
-    public class SwitchBranchInfo
-    {
-        public SwitchViewModel? Switch { get; set; }
-        public List<TrackViewModel>? Candidates { get; set; }
-        public Action<TrackViewModel?>? Callback { get; set; }
-    }
-
-    public class SignalViewModel : BaseElementViewModel
-    {
-        public override string TypeName => "Signal";
-        
-        // Collections for ComboBox
-        public System.Collections.Generic.List<string> AvailableTypes { get; } = new System.Collections.Generic.List<string> 
-        { 
-            "main", "distant", "repeater", "shunting" 
-        };
-
-        public System.Collections.Generic.List<string> AvailableFunctions { get; } = new System.Collections.Generic.List<string> 
-        { 
-            "exit", "home", "blocking" 
-        };
-
-        private string? _type;
-        public string? Type
-        {
-            get => _type;
-            set => SetProperty(ref _type, value);
-        }
-
-        private string? _function;
-        public string? Function
-        {
-            get => _function;
-            set => SetProperty(ref _function, value);
-        }
-        
-        private string? _relatedTrackId;
-        public string? RelatedTrackId
-        {
-            get => _relatedTrackId;
-            set => SetProperty(ref _relatedTrackId, value);
-        }
-
-        private double _pos;
-        public double Pos
-        {
-            get => _pos;
-            set => SetProperty(ref _pos, value);
-        }
-
-        public System.Collections.Generic.List<string> AvailableDirections { get; } = new System.Collections.Generic.List<string> { "up", "down" };
-        
-        private string _direction = "up";
-        public string Direction
-        {
-            get => _direction;
-            set 
-            {
-                if (SetProperty(ref _direction, value))
-                {
-                    OnPropertyChanged(nameof(IsFlipped));
-                }
-            }
-        }
-
-        public bool IsFlipped
-        {
-            get => _direction == "down";
-            set => Direction = value ? "down" : "up";
-        }
-    }
-
-
-    public class SwitchPositionViewModel : ObservableObject
-    {
-        private string? _switchRef;
-        public string? SwitchRef { get => _switchRef; set => SetProperty(ref _switchRef, value); }
-
-        private string _switchPosition = "normal";
-        public string SwitchPosition { get => _switchPosition; set => SetProperty(ref _switchPosition, value); }
-
-        public List<string> AvailablePositions { get; } = new() { "normal", "reverse" };
-
-        public ICommand? RemoveCommand { get; set; }
-    }
-
-    public class RouteViewModel : BaseElementViewModel
-    {
-        public override string TypeName => "Route";
-
-        private string _code = string.Empty;
-        public string Code { get => _code; set => SetProperty(ref _code, value); }
-
-        private string _description = string.Empty;
-        public string Description { get => _description; set => SetProperty(ref _description, value); }
-
-        private string _approachPointRef = string.Empty;
-        public string ApproachPointRef { get => _approachPointRef; set => SetProperty(ref _approachPointRef, value); }
-
-        private string _entryRef = string.Empty;
-        public string EntryRef { get => _entryRef; set => SetProperty(ref _entryRef, value); }
-
-        private string _exitRef = string.Empty;
-        public string ExitRef { get => _exitRef; set => SetProperty(ref _exitRef, value); }
-
-        private string _overlapEndRef = string.Empty;
-        public string OverlapEndRef { get => _overlapEndRef; set => SetProperty(ref _overlapEndRef, value); }
-
-        private string _proceedSpeed = "R";
-        public string ProceedSpeed { get => _proceedSpeed; set => SetProperty(ref _proceedSpeed, value); }
-
-        private bool _releaseTriggerHead;
-        public bool ReleaseTriggerHead { get => _releaseTriggerHead; set => SetProperty(ref _releaseTriggerHead, value); }
-
-        private string _releaseTriggerRef = string.Empty;
-        public string ReleaseTriggerRef { get => _releaseTriggerRef; set => SetProperty(ref _releaseTriggerRef, value); }
-
-        public ObservableCollection<SwitchPositionViewModel> SwitchAndPositions { get; } = new();
-        public ObservableCollection<SwitchPositionViewModel> OverlapSwitchAndPositions { get; } = new();
-        public ObservableCollection<ReleaseSectionViewModel> ReleaseSections { get; } = new();
-
-        public ICommand AddSwitchPositionCommand { get; }
-        public ICommand AddOverlapSwitchPositionCommand { get; }
-        public ICommand AddReleaseSectionCommand { get; }
-
-        public List<string> AvailableProceedSpeeds { get; } = new() { "R", "YY", "Y", "YG", "G" };
-
-        public RouteViewModel()
-        {
-            AddSwitchPositionCommand = new RelayCommand(_ => SwitchAndPositions.Add(new SwitchPositionViewModel { RemoveCommand = new RelayCommand(p => { if (p is SwitchPositionViewModel vm) SwitchAndPositions.Remove(vm); }) }));
-            AddOverlapSwitchPositionCommand = new RelayCommand(_ => OverlapSwitchAndPositions.Add(new SwitchPositionViewModel { RemoveCommand = new RelayCommand(p => { if (p is SwitchPositionViewModel vm) OverlapSwitchAndPositions.Remove(vm); }) }));
-            AddReleaseSectionCommand = new RelayCommand(_ => ReleaseSections.Add(new ReleaseSectionViewModel { RemoveCommand = new RelayCommand(p => { if (p is ReleaseSectionViewModel vm) ReleaseSections.Remove(vm); }) }));
-        }
-
-        public override double X { get => 0; set { } }
-        public override double Y { get => 0; set { } }
-    }
-
-    public class ReleaseSectionViewModel : ObservableObject
-    {
-        private string? _trackRef;
-        public string? TrackRef { get => _trackRef; set => SetProperty(ref _trackRef, value); }
-
-        private bool _flankProtection;
-        public bool FlankProtection { get => _flankProtection; set => SetProperty(ref _flankProtection, value); }
-
-        public ICommand? RemoveCommand { get; set; }
-
-        public List<bool> AvailableProtections { get; } = new() { true, false };
-    }
-
-    public class InfrastructureViewModel : BaseElementViewModel
-    {
-        public override string TypeName => "Infrastructure";
-        public ObservableCollection<CategoryViewModel> Categories { get; } = new();
-
-        public override double X { get => 0; set { } }
-        public override double Y { get => 0; set { } }
-    }
-
-    public class CategoryViewModel : ObservableObject
-    {
-        public string? Title { get; set; }
-        public ObservableCollection<BaseElementViewModel> Items { get; } = new ObservableCollection<BaseElementViewModel>();
-    }
-
     public class MainViewModel : ObservableObject
     {
         public ObservableCollection<BaseElementViewModel> Elements { get; } = new ObservableCollection<BaseElementViewModel>();
@@ -1554,3 +892,4 @@ namespace RailmlEditor.ViewModels
         }
     }
 }
+
