+++ RailmlEditor/Views/ExplorerView.xaml.cs
@@ -3,6 +3,7 @@ using System.Windows;
 using System.Windows.Controls;
 using System.Windows.Input;
 using RailmlEditor.ViewModels;
+using RailmlEditor.ViewModels.Elements;
 
 namespace RailmlEditor.Views
 {
@@ -99,3 +100,5 @@ namespace RailmlEditor.Views
         }
     }
 }
+
+
